사각형 색 구현

+++ WinAPI_AStar/WinAPI_AStar/AStar.h
@@ -27,7 +27,7 @@ namespace MomDra
 
 			}
 
-			double F() const noexcept
+			int F() const noexcept
 			{
 				return g + h;
 			}
